pyproject.toml
[tool.poetry]
name = "flatter"
version = "0.1.0"
description = ""
authors = ["Eduardo <eduardooa1980@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"underscore.py" = "^0.1.6"

[tool.poetry.dev-dependencies]
nose2 = "^0.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
